pyproject.toml
[project]
name = "langgraph-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg>=1.4",
    "ipykernel",
    "jupyter>=1.1.1",
    "langchain-core>=0.3.75",
    "langchain-deepseek>=0.1.4",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langsmith>=0.4.25",
    "matplotlib>=3.10.6",
    "notebook>=7.4.5",
    "openai>=1.107.0",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
]
